Created menu_options.py file which holds functions that return query results
